Fix: check column exists before removing in RemoveSorceryColumnsFromUsers migration

+++ db/migrate/20250603122214_remove_sorcery_columns_from_users.rb
@@ -1,8 +1,8 @@
 class RemoveSorceryColumnsFromUsers < ActiveRecord::Migration[8.0]
   def change
-    remove_column :users, :crypted_password, :string
-    remove_column :users, :salt, :string
-    remove_column :users, :reset_password_token, :string
-    remove_column :users, :password_reset_sent_at, :datetime
+    remove_column :users, :crypted_password, :string if column_exists?(:users, :crypted_password)
+    remove_column :users, :salt, :string if column_exists?(:users, :salt)
+    remove_column :users, :reset_password_token, :string if column_exists?(:users, :reset_password_token)
+    remove_column :users, :password_reset_sent_at, :datetime if column_exists?(:users, :password_reset_sent_at)
   end
 end
